content_root/samples/common/requirements.txt
bert-tensorflow
boto3
category_encoders
flake8
flask
Jinja2
joblib
lib/fiddler_utils-1.0.tar.gz
lightgbm
lime
numpy
pandas
pmlb
pyarrow
pytest
pyyaml
requests
scikit-learn==0.21.2
scipy
shap
tensorflow-gpu==1.14
tensorflow_datasets
unidecode
xgboost
